Module 7 add comment

+++ MainApp/views.py
@@ -1,7 +1,9 @@
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 from django.shortcuts import render, redirect
-from MainApp.models import Snippet
-from MainApp.forms import SnippetForm, UserRegistrationForm
 from django.contrib import auth
+from MainApp.models import Snippet
+from MainApp.forms import SnippetForm, UserRegistrationForm, CommentForm
 
 
 def index_page(request):
@@ -10,14 +12,14 @@ def index_page(request):
 
 
 def add_snippet(request):
-    if request.method == "GET":
+    if request.method == "GET":  # получить страницу с формой
         form = SnippetForm()
         context = {
             'form': form,
             'pagename': 'Добавление нового сниппета'
         }
         return render(request, 'pages/add_snippet.html', context)
-    elif request.method == "POST":
+    elif request.method == "POST":  # получить данные от формы
         form = SnippetForm(request.POST)
         if form.is_valid():
             snippet = form.save(commit=False)
@@ -26,8 +28,10 @@ def add_snippet(request):
         return redirect('snippets-list')
 
 
-def snippet_delete(request, snippet_id):
+def snippet_delete(requests, snippet_id):
     snippet = Snippet.objects.get(id=snippet_id)
+    if snippet.user != requests.user:
+        raise PermissionDenied()
     snippet.delete()
     return redirect('snippets-list')
 
@@ -51,11 +55,23 @@ def snippets_list(request):
 
 def snippet_detail(request, snippet_id):
     snippet = Snippet.objects.get(id=snippet_id)
+    comment_form = CommentForm()
     context = {
         'pagename': 'Информация о сниппете',
-        "snippet": snippet
+        "snippet": snippet,
+        "comment_form": comment_form
     }
     return render(request, 'pages/snippet_detail.html', context)
+
+
+@login_required
+def snippets_my(request):
+    my_snippets = Snippet.objects.filter(user=request.user)
+    context = {
+        'pagename': 'Мои сниппеты',
+        "snippets": my_snippets
+    }
+    return render(request, 'pages/view_snippets.html', context)
 
 
 def login_page(request):
@@ -80,19 +96,31 @@ def registration(request):
     if request.method == "GET":
         form = UserRegistrationForm()
         context = {
-            'form': form,
-            'pagename': 'Регистрация нового пользователя'
+            'pagename': 'Регистрация',
+            "form": form
         }
         return render(request, 'pages/registration.html', context)
-    if request.method == "POST":
+    elif request.method == "POST":
         form = UserRegistrationForm(request.POST)
         if form.is_valid():
             form.save()
-        else:
+        else:  # данные не валидные
             context = {
-                'form': form,
-                'pagename': 'Регистрация нового пользователя'
+                'pagename': 'Регистрация',
+                "form": form
             }
             return render(request, 'pages/registration.html', context)
         return redirect('home')
 
+
+def comment_create(request):
+    if request.method == "POST":
+        comment_form = CommentForm(request.POST)
+        snippet_id = request.POST.get("snippet_id")
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.author = request.user
+            snippet = Snippet.objects.get(id=snippet_id)
+            comment.snippet = snippet
+            comment.save()
+        return redirect(request.META.get('HTTP_REFERER', '/'))
